feat: Separate helper and get id from constructor
