+++ personalization-sdk/src/rees46/kotlin/com/rees46/sdk/REES46.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import com.personalization.BuildConfig
 import com.personalization.SDK
 import com.personalization.notification.core.NotificationHelper
+import com.personalization.notification.helpers.NotificationImageHelper
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
@@ -62,10 +63,11 @@ class REES46 private constructor() : SDK() {
             sdk.setOnMessageListener { data: Map<String, String> ->
                 coroutineScope.launch {
                     val images = withContext(Dispatchers.IO) {
-                        NotificationHelper.loadBitmaps(urls = data[NotificationHelper.NOTIFICATION_IMAGES])
+                        NotificationImageHelper.loadBitmaps(urls = data[NotificationHelper.NOTIFICATION_IMAGES])
                     }
                     NotificationHelper.createNotification(
                         context = context,
+                        notificationId = data.hashCode(),
                         data = data,
                         images = images,
                         currentIndex = 0
